requirements.txt
Django==1.7.6
httplib2==0.8
restclient==0.11.0
feedparser==5.1.3
Markdown==2.6.1
smartypants==1.6.0.3
uuid==1.30
psycopg2==2.5.4
Pillow==2.7.0
nose==1.3.4
versiontools==1.9.1
statsd==2.0.2
pep8==1.4.6
pyflakes==0.7.3
mccabe==0.2.1
flake8==2.0
tornado==3.1
BeautifulSoup==3.2.1
cssselect==0.8
lxml==3.2.3
fuzzywuzzy==0.2
sure==1.2.2
ipython==0.13.2
ipdb==0.7
rdflib==3.1.0
selenium==2.20.0
coverage==3.7.1
requests==0.13.3
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
six==1.3.0
factory_boy==2.1.1
python-ldap==2.4.13

djangowind==0.13.2
sorl==3.1
tagging==0.3-pre
typogrify==2.0.0
django-staticmedia==0.2.2
django-indexer==0.3.0
django-templatetag-sugar==0.1
django-paging==0.2.5
django-annoying==0.8.0
django-nose==1.2
django-appconf==0.6
django_compressor==1.4
django-statsd-mozilla==0.3.14
raven==5.1.1
django-bootstrap-form==2.0.3
sqlparse==0.1.11
django-debug-toolbar==1.1
django-waffle==0.10.1
django-jenkins==0.16.3
django-smoketest==0.3.0
django-extensions==1.2.5
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==0.9
django-treebeard==2.0b2
django-pagetree==1.0.9
django-pageblocks==1.0.0
django-quizblock==1.0.6
django-markwhat==2014.9.20
django-simple-captcha==0.4.1

